+++ word-map/src/main/java/com/margot/word_map/dto/request/UpdateLetterRequest.java
@@ -23,9 +23,9 @@ public class UpdateLetterRequest {
 
     @Min(1) @Max(10)
     @Schema(description = "Множитель бонуса за букву", example = "1")
-    private Short multiplier;
+    private Integer multiplier;
 
     @Min(1) @Max(50)
     @Schema(description = "Вес буквы (частота выпадения)", example = "6")
-    private Short weight;
+    private Integer weight;
 }
